Fix for devices before API11.
In API11 a Split touch events were added, so there is no problem with
handling overlapping views.
In pre-Honeycomb there is some problem with changing the order of
dispatching touch events after show/hide the drawer. This workaround
allows main view to handle the touch events if drawer is closed (and
not trying to be opened).

Follow the entry about API11 at
http://developer.android.com/about/versions/android-3.0.html for more
details.